Convert the Docker image to use Debian Buster as the base
Switch the base image of the Docker container created for this project
from "python:3.8-alpine" (Alpine Linux 3.11 with Python 3.8) to
"python:3.8-slim-buster" (Debian Linux "Buster" with Python 3.8).

This results in the ability to use pre-compiled binary wheels from PyPI
for the majority of the Python dependencies. This makes it much quicker
to build this image.

To keep the image size down to reasonable levels and to avoid installing
all sorts of unnecessary code and binaries into our container, we use
the multi-container and virtualenv technique described in this series of
articles:

https://pythonspeed.com/docker/
https://pythonspeed.com/articles/base-image-python-docker-images/
https://pythonspeed.com/articles/smaller-python-docker-images/
https://pythonspeed.com/articles/multi-stage-docker-python/